newBranch/requirements_standalone.txt
# AnComicsViewer - Standalone Application
# Core dependencies for PDF comic book viewing with YOLO panel/balloon detection

# GUI Framework
PySide6>=6.5.0

# PDF Processing  
PyMuPDF>=1.24.0

# Machine Learning
ultralytics>=8.2.0
torch>=2.0.0  # Required by ultralytics
torchvision>=0.15.0  # Required by ultralytics

# Scientific Computing
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# Optional: For enhanced performance
Pillow>=9.0.0  # Image processing
opencv-python>=4.8.0  # Computer vision operations

# Development (optional)
# pytest>=7.0.0  # For testing
# black>=23.0.0   # Code formatting
